Fix https://issues.apache.org/bugzilla/show_bug.cgi?id=56710 Do not perform wrapper mapping if context is being reloaded, as such mapping is useless and may fail unexpectedly. Let mapper know that context is being reloaded.
Ignore removeWrapper(), removeWelcomeFile() calls when context is being reloaded,
as that is useless work.
During undeploy the context should have already been unregistered.
During reload the context will be re-registered via addContextVersion().

It is backport of r1610220

git-svn-id: https://svn.apache.org/repos/asf/tomcat/tc7.0.x/trunk@1611472 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/catalina/connector/MapperListener.java
@@ -391,11 +391,6 @@ public class MapperListener extends LifecycleMBeanBase
      */
     private void unregisterContext(Context context) {
 
-        // Don't un-map a context that is paused
-        if (context.getPaused()){
-            return;
-        }
-
         String contextPath = context.getPath();
         if ("/".equals(contextPath)) {
             contextPath = "";
@@ -407,8 +402,23 @@ public class MapperListener extends LifecycleMBeanBase
                     contextPath, connector));
         }
 
-        mapper.removeContextVersion(hostName, contextPath,
-                context.getWebappVersion());
+        if (context.getPaused()) {
+            if (log.isDebugEnabled()) {
+                log.debug(sm.getString("mapperListener.pauseContext",
+                        contextPath, connector));
+            }
+
+            mapper.pauseContextVersion(context, hostName, contextPath,
+                    context.getWebappVersion());
+        } else {
+            if (log.isDebugEnabled()) {
+                log.debug(sm.getString("mapperListener.unregisterContext",
+                        contextPath, connector));
+            }
+
+            mapper.removeContextVersion(hostName, contextPath,
+                    context.getWebappVersion());
+        }
     }
 
 
@@ -482,12 +492,7 @@ public class MapperListener extends LifecycleMBeanBase
             if (obj instanceof Wrapper) {
                 unregisterWrapper((Wrapper) obj);
             } else if (obj instanceof Context) {
-                Context c = (Context) obj;
-                // Only unregister if not paused. If paused, need to keep
-                // registration in place to prevent 404's during reload
-                if (!c.getPaused()) {
-                    unregisterContext(c);
-                }
+                unregisterContext((Context) obj);
             } else if (obj instanceof Host) {
                 unregisterHost((Host) obj);
             }
